Move SelectItems and SelectItemsAlt to their own component
