+++ client/landing/stepper/declarative-flow/internals/steps-repository/intent-step/style.scss
@@ -6,21 +6,21 @@ $gray-100: #101517;
 $gray-60: #50575e;
 
 .intent-step {
-	.intent-screen-select-items__item {
-		.intent-screen-select-items__item-title {
+	.select-items__item {
+		.select-items__item-title {
 			font-weight: 400;
 			margin: 0;
 		}
 
-		button.intent-screen-select-items__item-button {
+		button.select-items__item-button {
 			box-shadow: rgba( 0, 0, 0, 0.05 ) 0 1px 2px 0;
 			border-radius: 4px;
 			border: 1px solid rgb( 195, 196, 199 );
 		}
 	}
 
-	.intent-screen-select-items-alt__item {
-		.intent-screen-select-items-alt__item-description {
+	.select-items-alt__item {
+		.select-items-alt__item-description {
 			margin: 0;
 		}
 	}
