+++ packages/onboarding/src/index.ts
@@ -11,7 +11,11 @@ export { default as ScrollToTop } from './scroll-to-top';
 export { default as Progress } from './progress';
 export { default as Hooray } from './hooray';
 export { default as Confetti } from './confetti';
-export * from './intent-screen';
+export { default as IntentScreen } from './intent-screen';
+export { default as SelectItems } from './select-items';
+export { default as SelectItemsAlt } from './select-items-alt';
 export { default as MShotsImage } from './mshots-image';
 export { default as Notice } from './notice';
+export type { SelectItem } from './select-items';
+export type { SelectItemAlt } from './select-items-alt';
 export type { MShotsOptions } from './mshots-image';
