--- packages/onboarding/src/intent-screen/index.ts
@@ -1,4 +1,0 @@
-export { default as IntentScreen } from './components/intent-screen';
-export { default as SelectItems } from './components/intent-screen-select-items';
-export { default as SelectItemsAlt } from './components/intent-screen-select-items-alt';
-export * from './types';
