+++ packages/onboarding/src/select-items/style.scss
@@ -2,22 +2,25 @@
 @import '@wordpress/base-styles/_mixins.scss';
 @import '@automattic/typography/styles/fonts';
 
-.intent-screen-select-items-alt {
+.select-items {
 	display: flex;
 	flex-direction: column;
 	justify-content: center;
+	margin-top: -12px;
+	margin-bottom: 16px;
 	padding-left: 40px;
 
 	@include break-small {
+		margin-bottom: 40px;
 		padding-left: 60px;
 	}
 }
 
-.intent-screen-select-items-alt__item {
+.select-items__item {
 	display: flex;
 	align-items: flex-start;
 	position: relative;
-	padding: 8px 0;
+	padding: 24px 0;
 
 	@include break-small {
 		width: 342px;
@@ -25,6 +28,15 @@
 
 	@include break-medium {
 		width: 456px;
+	}
+
+	&-icon {
+		position: absolute;
+		top: 24px;
+		left: -40px;
+		@include break-small {
+			left: -48px;
+		}
 	}
 
 	&-info-wrapper {
@@ -37,35 +49,42 @@
 
 	&-info {
 		flex: 1;
-		padding-right: 6px;
+
+		@include break-mobile {
+			padding-right: 20px;
+		}
+	}
+
+	&-title {
+		margin-bottom: 8px;
+		font-size: $font-title-small;
+		line-height: 1.2;
+		color: #101517;
 	}
 
 	&-description {
 		font-size: $font-body-small;
 		color: #646970;
-		margin-bottom: 0;
+
+		> p:last-child {
+			margin-bottom: 0;
+		}
 	}
 
-	&-button {
-		margin-top: 8px;
-		padding: 0;
+	button.select-items__item-button {
+		margin-top: 24px;
+		border-radius: 4px; /* stylelint-disable-line */
 		min-width: 130px;
-		border: none;
-		background: none;
+		box-shadow: 0 1px 2px rgba( 0, 0, 0, 0.05 );
 		color: #101517;
-		text-decoration: underline;
 		font-weight: 500; /* stylelint-disable-line scales/font-weights */
 
 		@include break-mobile {
 			margin-top: 0;
 		}
-
-		&:not( [disabled] ):hover {
-			color: var( --color-neutral-70 );
-		}
 	}
 
-	&-disabled-info svg {
-		fill: var( --color-neutral-20 );
+	& + & {
+		border-top: 1px solid rgba( 220, 220, 222, 0.64 );
 	}
 }
